+++ src/Symfony/Component/JsonStreamer/JsonStreamWriter.php
@@ -18,9 +18,11 @@ use Symfony\Component\JsonStreamer\Mapping\GenericTypePropertyMetadataLoader;
 use Symfony\Component\JsonStreamer\Mapping\PropertyMetadataLoader;
 use Symfony\Component\JsonStreamer\Mapping\PropertyMetadataLoaderInterface;
 use Symfony\Component\JsonStreamer\Mapping\Write\AttributePropertyMetadataLoader;
+use Symfony\Component\JsonStreamer\Transformer\BcMathNumberValueObjectTransformer;
 use Symfony\Component\JsonStreamer\Transformer\DateIntervalValueObjectTransformer;
 use Symfony\Component\JsonStreamer\Transformer\DateTimeValueObjectTransformer;
 use Symfony\Component\JsonStreamer\Transformer\DateTimeZoneValueObjectTransformer;
+use Symfony\Component\JsonStreamer\Transformer\GmpNumberValueObjectTransformer;
 use Symfony\Component\JsonStreamer\Transformer\PropertyValueTransformerInterface;
 use Symfony\Component\JsonStreamer\Transformer\ValueObjectTransformerInterface;
 use Symfony\Component\JsonStreamer\Write\StreamWriterGenerator;
@@ -110,6 +112,8 @@ final class JsonStreamWriter implements StreamWriterInterface
             \DateTimeInterface::class => new DateTimeValueObjectTransformer(),
             \DateInterval::class => new DateIntervalValueObjectTransformer(),
             \DateTimeZone::class => new DateTimeZoneValueObjectTransformer(),
+            \BcMath\Number::class => new BcMathNumberValueObjectTransformer(),
+            \GMP::class => new GmpNumberValueObjectTransformer(),
         ];
 
         $transformersContainer = new class($transformers) implements ContainerInterface {