+++ src/Symfony/Component/JsonStreamer/Tests/Fixtures/Model/DummyWithGmpNumber.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symfony\Component\JsonStreamer\Tests\Fixtures\Model;
+
+class DummyWithGmpNumber
+{
+    public \GMP $gmp;
+}