+++ src/Symfony/Component/JsonStreamer/Tests/JsonStreamReaderTest.php
@@ -11,16 +11,19 @@
 
 namespace Symfony\Component\JsonStreamer\Tests;
 
+use PHPUnit\Framework\Attributes\RequiresPhpExtension;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\JsonStreamer\JsonStreamReader;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Enum\DummyBackedEnum;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Mapping\SyntheticPropertyMetadataLoader;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\ClassicDummy;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithBcMathNumber;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithDateIntervals;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithDateTimes;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithDateTimeZones;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithGenerics;
+use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithGmpNumber;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithNameAttributes;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithNullableProperties;
 use Symfony\Component\JsonStreamer\Tests\Fixtures\Model\DummyWithPhpDoc;
@@ -220,6 +223,28 @@ class JsonStreamReaderTest extends TestCase
             $this->assertInstanceOf(DummyWithDateTimeZones::class, $read);
             $this->assertEquals(new \DateTimeZone('Asia/Tokyo'), $read->timezone);
         }, '{"timezone":"Asia/Tokyo"}', Type::object(DummyWithDateTimeZones::class));
+    }
+
+    #[RequiresPhpExtension('bcmath')]
+    public function testReadObjectWithBcMathNumber()
+    {
+        $reader = JsonStreamReader::create([], $this->streamReadersDir);
+
+        $this->assertRead($reader, function (mixed $read) {
+            $this->assertInstanceOf(DummyWithBcMathNumber::class, $read);
+            $this->assertEquals(new \BcMath\Number('3.14'), $read->number);
+        }, '{"number":"3.14"}', Type::object(DummyWithBcMathNumber::class));
+    }
+
+    #[RequiresPhpExtension('gmp')]
+    public function testReadObjectWithGmpNumber()
+    {
+        $reader = JsonStreamReader::create([], $this->streamReadersDir);
+
+        $this->assertRead($reader, function (mixed $read) {
+            $this->assertInstanceOf(DummyWithGmpNumber::class, $read);
+            $this->assertEquals(new \GMP('99999999999999999999'), $read->gmp);
+        }, '{"gmp":"99999999999999999999"}', Type::object(DummyWithGmpNumber::class));
     }
 
     public function testReadUnion()